refactor: widen instrument param ids to u16
PolySynth's flat id space had 11 ids left (23, 30, 31, 246-253):
param_specs runs to 193, the FX reorder handles fill the gaps, and
the custom-LFO breakpoint block owns 195-245. Every LFO/matrix
feature from here needs more room than that.

Event.set_param and DrumMachine were already u16 and everything
else truncated on the way in, so this only removes the odd ones
out - PolySynth, Sampler, SoundfontPlayer, Pad, and the editor
cursor/automation/undo plumbing that feeds them. Existing ids keep
their values (dest_pitch/dest_amp stay 254/255), so saved projects
and presets load unchanged.

Drops the @truncate at PolySynth's event boundary and the
`id <= 0xFF` guards in undo's liveParamValue, both of which would
have silently swallowed any id past 255.

+++ src/dsp/device.zig
@@ -25,11 +25,11 @@ pub const Event = union(enum) {
     pitch_bend: struct { bend: i16 },
     /// Nudge editor parameter `id` by `steps` (signed). Applied on the audio
     /// thread so UI edits never race the reader - see PolySynth.adjustParam.
-    /// `id` is u16, not u8: DrumMachine.paramId packs a pad index (up to 64)
-    /// into the high bits, which no longer fits u8 now that pad count grew
-    /// past 15. Every other device's own adjustParam/setParamAbsolute still
-    /// takes a plain u8 (their param counts are all well under 256) - the
-    /// wider id only matters at this shared event boundary.
+    /// `id` is u16 everywhere: DrumMachine.paramId packs a pad index (up to
+    /// 64) into the high bits, and PolySynth's own flat id space ran out of
+    /// u8 room once the per-LFO sync/retrigger/slew blocks landed. Every
+    /// device's adjustParam/setParamAbsolute takes the same u16, so no
+    /// truncation happens anywhere between the editor and the device.
     set_param: struct { id: u16, steps: i32 },
     /// Set editor parameter `id` to an absolute value - the counterpart to
     /// `set_param` automation curves need, since a curve knows the value it
@@ -78,7 +78,7 @@ pub const Event = union(enum) {
 /// editor can look either table up through one shared type regardless of
 /// which instrument the current track holds.
 pub const AutomatableParam = struct {
-    id: u8,
+    id: u16,
     label: []const u8,
     section: []const u8,
     range: [2]f32,
